requirements.txt
absl-py==1.4.0
adal==1.2.7
aiohttp==3.8.3
aiohttp-cors==0.7.0
aioredis==1.3.1
aiorwlock==1.3.0
aiosignal==1.3.1
alembic==1.8.1
analytics-python==1.4.0
ansimarkup==1.5.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argparse==1.4.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
avro==1.11.1
azure-core==1.26.2
azure-identity==1.12.0
azure-storage-blob==12.9.0
azure-storage-file-share==12.7.0
backcall==0.2.0
backoff==1.10.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
blessed==1.19.1
boto3==1.21.0
botocore==1.24.21
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-params==0.3.0
cloudevents==1.9.0
cloudpickle==2.2.1
colorama==0.4.6
colorful==0.5.5
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
coverage==7.1.0
cryptography==39.0.0
cycler==0.11.0
databricks-cli==0.17.4
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
distro==1.8.0
docker==6.0.1
docstring-parser==0.15
duty==0.7.0
ecdsa==0.18.0
entrypoints==0.4
enum-compat==0.0.3
evidently==0.1.54.dev0
executing==1.2.0
failprint==0.8.0
fastapi==0.75.2
fastapi-utils==0.2.1
fastjsonschema==2.16.2
filelock==3.9.0
fire==0.5.0
flake8==3.9.2
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.5.3
gcsfs==2023.1.0
gitdb==4.0.10
gitpython==3.1.30
google-api-core==2.11.0
google-api-python-client==1.12.11
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.13.1
google-cloud-bigquery==2.34.3
google-cloud-build==3.11.1
google-cloud-core==2.3.2
google-cloud-functions==1.10.1
google-cloud-resource-manager==1.8.1
google-cloud-scheduler==2.9.1
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gpustat==1.0.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
hiredis==2.1.1
httplib2==0.19.1
httptools==0.5.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
jupyter-client==8.0.2
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.2.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaggle==1.5.12
keras==2.8.0
keras-preprocessing==1.1.2
kfp==1.8.16
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kiwisolver==1.4.4
kserve==0.9.0
kubernetes==18.20.0
libclang==15.0.6.1
Mako==1.2.4
markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.6.1
minio==7.1.3
mistune==2.0.4
mlflow==1.25.1
mlserver==1.0.0
mlserver-mlflow==1.0.0
monotonic==1.6
msal==1.21.0
msal-extensions==1.0.0
msgpack==1.0.4
msrest==0.7.1
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.21.6
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
pluggy==1.0.0
portalocker==2.7.0
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-grpc-prometheus==0.7.0
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.9.2
pyflakes==2.3.1
pygments==2.14.0
PyJWT==2.6.0
pymysql==1.0.2
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-json-logger==2.0.4
python-multipart==0.0.5
python-rapidjson==1.9
python-slugify==8.0.0
python-terraform==0.10.1
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==25.0.0
querystring-parser==1.2.4
ray==1.10.0
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
s3transfer==0.5.2
scikit-learn==1.2.1
scipy==1.8.1
Send2Trash==1.8.0
setuptools==67.1.0
shapely==1.8.5.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlalchemy==1.4.41
sqlalchemy-utils==0.38.3
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.17.1
starlette-exporter==0.15.1
statsmodels==0.13.5
strip-hints==0.1.10
table-logger==0.3.6
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io==0.24.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch-model-archiver==0.7.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
tritonclient==2.18.0
typer==0.7.0
typing-extensions==4.4.0
uritemplate==3.0.1
urllib3==1.26.14
uvicorn==0.17.6
uvloop==0.17.0
validators==0.18.2
watchgod==0.8.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.0
websockets==10.4
werkzeug==2.2.2
wheel==0.38.4
widgetsnbextension==3.6.1
wrapt==1.14.1
yarl==1.8.2
zenml==0.32.1
zipp==3.12.0
